Update workflow in docs
It is now recommended to fetch cells into a separate directory for the downstream analysis. By design we only support fetching of one dataset per-directory.

In future pipeline_integration and pipeline_export will be moved to a new scxl repo along with with a streamlined version of pipeline_scxl from the repo tenx.