Standardise pipeline configuration
All pipelines are now configured with the same code block.

After doing "import cgatcore.pipeline as P" it is necessary to
set "P.parameters.HAVE_INITIALIZED = False" to avoid an issue
with sphinx autodocs. In brief autodocs uses importlib.import_module
to import all the pipelines. Because modules are imported only
once in Python, "P.parameters.HAVE_INITIALIZED" will be set to
"True" by previous imported pipelines. If not corrected this causes
issues with missing keys or key types in the PARAMs dict and the
docs will fail to build.